Adjusted Review class

+++ app/src/main/java/com/mobdeve/s13/group5/grubrev/Review.kt
@@ -4,5 +4,5 @@ class Review(
     val restaurant : String,
     val user: String,
     val comment: String,
-    val rating: Int) {
+    val rating: Double) {
 }
